requirements.txt
matplotlib==3.7.4
nibabel==5.2.0
numpy==1.24.1
opencv_python==4.8.1.78
PyYAML==6.0.1
scipy==1.11.4
segment_anything==1.0
torch==2.1.1+cu118
tqdm==4.66.1
